Make a new VehicleStatus model
Move status and isUnassignedBySwiftly here.
