+++ assets/src/components/incomingBox.tsx
@@ -1,11 +1,7 @@
 import React, { useContext } from "react"
 import DispatchContext from "../contexts/dispatchContext"
-import {
-  directionOnLadder,
-  LadderVehicle,
-  status,
-  VehicleDirection,
-} from "../models/ladderVehicle"
+import { directionOnLadder, VehicleDirection } from "../models/ladderVehicle"
+import { status } from "../models/vehicleStatus"
 import { Vehicle, VehicleId } from "../skate"
 import { selectVehicle } from "../state"
 import { LadderDirection } from "./ladder"
@@ -30,9 +26,7 @@ const IncomingBoxVehicle = ({
 
   return (
     <button
-      className={`m-incoming-box__vehicle ${status({
-        vehicle,
-      } as LadderVehicle)} ${selectedClass}`}
+      className={`m-incoming-box__vehicle ${status(vehicle)} ${selectedClass}`}
       onClick={() => dispatch(selectVehicle(vehicle.id))}
     >
       <VehicleIcon
