+++ assets/tests/models/ladderVehicle.test.ts
@@ -4,13 +4,11 @@ import {
   byDirectionAndY,
   directionOnLadder,
   firstOpenLane,
-  isUnassignedBySwiftly,
   LadderVehicle,
   putIntoLanes,
-  status,
   VehicleDirection,
 } from "../../src/models/ladderVehicle"
-import { DataDiscrepancy, Vehicle } from "../../src/skate"
+import { Vehicle } from "../../src/skate"
 
 describe("directionOnLadder", () => {
   test("determines the vehicle direction relative the ladder direction", () => {
@@ -68,113 +66,6 @@ describe("putIntoLanes", () => {
 
     expect(ys).toEqual(expectedYs)
     expect(lanes).toEqual(expectedLanes)
-  })
-})
-
-describe("status", () => {
-  test("returns 'off-course' if isUnassignedBySwiftly", () => {
-    const vehicle: Vehicle = {
-      dataDiscrepancies: [
-        {
-          attribute: "trip_id",
-          sources: [
-            {
-              id: "swiftly",
-              value: null,
-            },
-            {
-              id: "busloc",
-              value: "busloc-trip-id",
-            },
-          ],
-        },
-      ],
-    } as Vehicle
-    const ladderVehicle: LadderVehicle = { vehicle } as LadderVehicle
-
-    expect(status(ladderVehicle)).toEqual("off-course")
-  })
-
-  test("returns the vehicle's schedule adherence status otherwise", () => {
-    const scheduleAdherenceStatus = "on-time"
-    const vehicle: Vehicle = {
-      scheduleAdherenceStatus,
-      dataDiscrepancies: [] as DataDiscrepancy[],
-    } as Vehicle
-    const ladderVehicle: LadderVehicle = { vehicle } as LadderVehicle
-
-    expect(status(ladderVehicle)).toEqual("on-time")
-  })
-})
-
-describe("isUnassignedBySwiftly", () => {
-  test("returns true if there is a trip_id data discrepancy where swiftly is null and busloc has a value", () => {
-    const vehicle: Vehicle = {
-      dataDiscrepancies: [
-        {
-          attribute: "trip_id",
-          sources: [
-            {
-              id: "swiftly",
-              value: null,
-            },
-            {
-              id: "busloc",
-              value: "busloc-trip-id",
-            },
-          ],
-        },
-      ],
-    } as Vehicle
-    const ladderVehicle: LadderVehicle = { vehicle } as LadderVehicle
-
-    expect(isUnassignedBySwiftly(ladderVehicle)).toBeTruthy()
-  })
-
-  test("returns false if the swiftly defined a value", () => {
-    const vehicle: Vehicle = {
-      dataDiscrepancies: [
-        {
-          attribute: "trip_id",
-          sources: [
-            {
-              id: "swiftly",
-              value: "swiftly-trip-id",
-            },
-            {
-              id: "busloc",
-              value: "busloc-trip-id",
-            },
-          ],
-        },
-      ],
-    } as Vehicle
-    const ladderVehicle: LadderVehicle = { vehicle } as LadderVehicle
-
-    expect(isUnassignedBySwiftly(ladderVehicle)).toBeFalsy()
-  })
-
-  test("returns false if there isn't a trip_id data discrepancy", () => {
-    const vehicle: Vehicle = {
-      dataDiscrepancies: [
-        {
-          attribute: "route_id",
-          sources: [
-            {
-              id: "swiftly",
-              value: "swiftly-route-id",
-            },
-            {
-              id: "busloc",
-              value: "busloc-route-id",
-            },
-          ],
-        },
-      ],
-    } as Vehicle
-    const ladderVehicle: LadderVehicle = { vehicle } as LadderVehicle
-
-    expect(isUnassignedBySwiftly(ladderVehicle)).toBeFalsy()
   })
 })
